requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
python-dotenv==1.0.0
Jinja2==3.1.2
requests==2.31.0
Pillow==10.1.0
gunicorn==21.2.0